requirements.txt
langchain==0.0.351
langchain-community==0.0.4
openai==1.6.1
python-dotenv==1.0.1
streamlit==1.31.1
streamlit-chat==0.1.1